requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0
bleach==6.1.0
Flask-Limiter==3.5.0